requirements.txt
spotipy
ytmusicapi
inquirer
python-dotenv
